+++ packages/fxa-settings/src/pages/ResetPassword/ResetPassword/en.ftl
